refactor: jobs.py - make it more modular - datastore_utils.py
Start of major refactoring to make jobs.py smaller
